Use C++ OpenImageIO library to read images.
Changed OpenImageIO build flags to help linking.
By default, delete the existing CMake build files each run.

Added "ImageShared" struct to return an image between the Rust/C++ FFI.

Change use of floating point Bounding Box to integer based.

Add support for RGB or RGBA images.

+++ src/colorutils.rs
@@ -26,7 +26,7 @@ pub fn convert_linear_to_srgb(x: f32) -> f32 {
     }
 }
 
-pub fn apply_color_matrix_inplace(pixels: &mut [f32], num_channels: u8, matrix: na::Matrix4<f32>) {
+pub fn apply_color_matrix_inplace(pixels: &mut [f32], num_channels: i32, matrix: na::Matrix4<f32>) {
     let pixel_count = pixels.len() / (num_channels as usize);
     match num_channels {
         3 => {
